moved help functions to new folder services
